+++ src/sql_type.ts
@@ -7,12 +7,12 @@ export interface SqlCustomer {
   cust_phone: string;
   cust_email: string | null;
   cust_account: string | null;
-  cust_password: string;
+  // cust_password: string;
 }
 export interface SqlShop {
   shop_id: number;
   shop_name: string;
-  shop_password: string;
+  // shop_password: string;
   shop_location: string | null;
   shop_phone: string;
   delivery_range: string | null;
@@ -21,7 +21,7 @@ export interface SqlShop {
 export interface SqlCourier {
   cour_id: number;
   cour_name: string;
-  cour_password: string;
+  // cour_password: string;
   cour_phone: string;
   cour_living: string | null;
   cour_onboarding_time: Date;
@@ -36,6 +36,28 @@ export interface SqlDish {
   dish_score: number | null;
   dish_sales: number;
 }
-export interface SqlOrder {
-  
+export interface SqlOrder {}
+
+type EraseDate<T> = T extends Date ? string : T;
+type SqlToReply<T extends object> = {
+  [k in keyof T]: T[k] extends {}
+    ? EraseDate<T[k]>
+    : NonNullable<EraseDate<T[k]>> | undefined;
+};
+
+/**
+ * Convert NULL from sql to undefined
+ */
+export function sql2Reply<T extends object>(row: T): SqlToReply<T> {
+  const target: any = {};
+  for (const [k, v] of Object.entries(row)) {
+    if (v !== null) {
+      if (v instanceof Date) {
+        target[k] = v.toISOString();
+      } else {
+        target[k] = v;
+      }
+    }
+  }
+  return target;
 }